ts-config handling
* extract
* no-longer console.log, instead prepend context specific info to the errors message
